+++ PersistencePlugin/src/com/elmakers/mine/bukkit/plugins/persistence/PersistencePlugin.java
@@ -4,8 +4,11 @@ import java.io.File;
 import java.util.logging.Logger;
 
 import org.bukkit.Server;
+import org.bukkit.event.Event.Priority;
+import org.bukkit.event.Event.Type;
 import org.bukkit.plugin.PluginDescriptionFile;
 import org.bukkit.plugin.PluginLoader;
+import org.bukkit.plugin.PluginManager;
 import org.bukkit.plugin.java.JavaPlugin;
 
 public class PersistencePlugin extends JavaPlugin
@@ -50,6 +53,10 @@ public class PersistencePlugin extends JavaPlugin
 	@Override
 	public void onEnable()
 	{
+		PluginManager pm = getServer().getPluginManager();
+		
+        pm.registerEvent(Type.PLAYER_COMMAND, playerListener, Priority.Normal, this);
+        
 		PluginDescriptionFile pdfFile = this.getDescription();
         log.info(pdfFile.getName() + " version " + pdfFile.getVersion() + " is enabled");
 	}
@@ -64,6 +71,7 @@ public class PersistencePlugin extends JavaPlugin
 	 */
 	
 	private static PersistencePlugin pluginInstance = null;
+	private final PersistencePlayerListener playerListener = new PersistencePlayerListener();
 	private Persistence persistence = null;
 	private static final Logger log = Logger.getLogger("Minecraft");